generalize GenericRecipePage.mouseClicked
